Drillo/requirements.txt
asgiref==3.8.1
boto3==1.36.22
botocore==1.36.22
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cloudinary==1.41.0
cryptography==44.0.0
diff-match-patch==20241021
Django==5.1.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-dynamic-model==0.3.0
django-import-export==4.3.3
django-storages==1.14.4
idna==3.10
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.2
pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.11.2
six==1.17.0
sqlparse==0.5.2
tablib==3.7.0
urllib3==2.2.3
